+++ apps/kimi-code/src/feedback/codebase/index.ts
@@ -1,4 +1,3 @@
-export * from './attach';
 export * from './packager';
 export * from './scanner';
 export * from './types';
