+++ apps/kimi-code/src/feedback/codebase/scanner.ts
